Update unit-test project settings.

+++ iOS/tests/iOSUnitTests/pubnub/libs/PubNub/Network/Transport/PNConnection.m
@@ -1327,7 +1327,7 @@ void writeStreamCallback(CFWriteStreamRef stream, CFStreamEventType type, void *
     [PNLogger logConnectionInfoMessageFrom:self message:^NSString * {
 
         return [NSString stringWithFormat:@"[CONNECTION::%@] RETRY CONNECTION (%lu / %d) (STATE: %lu)",
-                self.name ? self.name : self, (unsigned long)self.connectionRetryCount, kPNMaximumConnectionRetryCount, self.state];
+                self.name ? self.name : self, (unsigned long)self.connectionRetryCount, (int)kPNMaximumConnectionRetryCount, self.state];
     }];
     
     // Check whether reconnection was issued because of SSL error or not
